Interger.compare(int, int) is introduced in Java 7, revert to Float.compare(float, float).
git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@1519 3a85a776-1598-11dc-ad04-adf9e166afdd

+++ SubDisc/src/nl/liacs/subdisc/Condition.java
@@ -395,7 +395,8 @@ public class Condition implements Comparable<Condition>
 					// else if (A.subset > B.subset) return 1;
 					// else if (same subset size) compare size
 					// else (same size) order sets return compare(A[0], B[0])
-					return Integer.compare(itsNominalValueSet.size(), theCondition.itsNominalValueSet.size());
+					//NOTE Integer.compare() is only introduced in Java 7
+					return Float.compare(itsNominalValueSet.size(), theCondition.itsNominalValueSet.size());
 				}
 			}
 			case NUMERIC :
